back-end/requirements.txt
pika

mysql-connector-python

